Profile Picture
This picture is used in the code within the project in "contact me" section.

Note: This picture showcasing my personal identity. Anyone with this code is only allowed to use this for portfolio purpose.
Remember it's a great effort to choose or showcase your own profile picture as it emphasizes strong verified impressions.